extract change_email_button
